refactor(layout): extract table row helpers
Move table row grouping, row geometry, and label-line merging into a focused helper module. This keeps the main table detector centered on candidate selection and table assembly without changing table detection behavior.
